server/requirements.txt
flask
flask-cors
gunicorn==20.1.0
google-cloud-storage==2.5.0
google-cloud-translate==3.12.0
google-cloud-datastore==2.11.0
singleton-decorator
openai==0.27.7
transformers==4.34.1
torch==2.1.0
spacy==3.7.2
numpy==1.26.4